Move and rename CORS related tests for improved readability
